requirements.txt
environs
GitPython
pyrogram
tgcrypto
uvloop
